refactored variables, declaration placed out of test functions
