+++ src/main/java/rueppellii/backend2/tribes/gameUtility/purchaseService/PurchaseService.java
@@ -3,17 +3,11 @@ package rueppellii.backend2.tribes.gameUtility.purchaseService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import rueppellii.backend2.tribes.building.persistence.model.Building;
-import rueppellii.backend2.tribes.building.service.BuildingService;
-import rueppellii.backend2.tribes.kingdom.exception.KingdomNotFoundException;
 import rueppellii.backend2.tribes.kingdom.persistence.model.Kingdom;
-import rueppellii.backend2.tribes.kingdom.service.KingdomService;
-import rueppellii.backend2.tribes.building.exception.BuildingNotFoundException;
 import rueppellii.backend2.tribes.resource.service.ResourceService;
-import rueppellii.backend2.tribes.resource.utility.ResourceType;
 import rueppellii.backend2.tribes.resource.exception.NoResourceException;
-import rueppellii.backend2.tribes.troop.exception.TroopNotFoundException;
-import rueppellii.backend2.tribes.troop.persistence.model.Troop;
-import rueppellii.backend2.tribes.troop.service.TroopService;
+import rueppellii.backend2.tribes.integrationTests.troop.persistence.model.Troop;
+import rueppellii.backend2.tribes.integrationTests.troop.service.TroopService;
 
 import java.util.List;
