+++ src/main/java/rueppellii/backend2/tribes/integrationTests/troop/exception/TroopNotFoundException.java
@@ -1,4 +1,4 @@
-package rueppellii.backend2.tribes.troop.exception;
+package rueppellii.backend2.tribes.integrationTests.troop.exception;
 
 public class TroopNotFoundException extends Exception {
     public TroopNotFoundException(String message) {
