+++ src/main/java/rueppellii/backend2/tribes/integrationTests/troop/persistence/repository/TroopRepository.java
@@ -1,9 +1,9 @@
-package rueppellii.backend2.tribes.troop.persistence.repository;
+package rueppellii.backend2.tribes.integrationTests.troop.persistence.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import rueppellii.backend2.tribes.kingdom.persistence.model.Kingdom;
-import rueppellii.backend2.tribes.troop.persistence.model.Troop;
+import rueppellii.backend2.tribes.integrationTests.troop.persistence.model.Troop;
 
 import java.util.List;
