+++ src/main/java/rueppellii/backend2/tribes/integrationTests/troop/utility/TroopFactory.java
@@ -1,0 +1,10 @@
+package rueppellii.backend2.tribes.integrationTests.troop.utility;
+
+import rueppellii.backend2.tribes.integrationTests.troop.persistence.model.Troop;
+
+public class TroopFactory {
+
+    public static Troop makeTroop(){
+        return new Troop();
+    }
+}
