+++ src/main/java/rueppellii/backend2/tribes/integrationTests/troop/web/TroopControllerAdvice.java
@@ -1,4 +1,4 @@
-package rueppellii.backend2.tribes.troop.web;
+package rueppellii.backend2.tribes.integrationTests.troop.web;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import rueppellii.backend2.tribes.building.exception.BuildingNotFoundException;
 import rueppellii.backend2.tribes.kingdom.exception.KingdomNotFoundException;
 import rueppellii.backend2.tribes.resource.exception.NoResourceException;
-import rueppellii.backend2.tribes.troop.exception.TroopNotFoundException;
+import rueppellii.backend2.tribes.integrationTests.troop.exception.TroopNotFoundException;
 import rueppellii.backend2.tribes.user.util.ErrorResponse;
 
 @ControllerAdvice(assignableTypes = TroopController.class)
