+++ src/main/java/rueppellii/backend2/tribes/kingdom/persistence/model/Kingdom.java
@@ -11,7 +11,7 @@ import rueppellii.backend2.tribes.location.persistence.model.Location;
 import rueppellii.backend2.tribes.resource.presistence.model.Resource;
 import rueppellii.backend2.tribes.progression.persistence.ProgressionModel;
 import rueppellii.backend2.tribes.resource.service.ResourceService;
-import rueppellii.backend2.tribes.troop.persistence.model.Troop;
+import rueppellii.backend2.tribes.integrationTests.troop.persistence.model.Troop;
 import rueppellii.backend2.tribes.user.persistence.model.ApplicationUser;
 
 import javax.persistence.*;
