+++ src/main/java/rueppellii/backend2/tribes/kingdom/web/KingdomController.java
@@ -10,7 +10,7 @@ import rueppellii.backend2.tribes.kingdom.service.KingdomService;
 import rueppellii.backend2.tribes.kingdom.exception.KingdomNotFoundException;
 import rueppellii.backend2.tribes.building.exception.BuildingNotFoundException;
 import rueppellii.backend2.tribes.kingdom.utility.KingdomNameDTO;
-import rueppellii.backend2.tribes.troop.exception.TroopNotFoundException;
+import rueppellii.backend2.tribes.integrationTests.troop.exception.TroopNotFoundException;
 
 import java.security.Principal;
