--- src/main/java/rueppellii/backend2/tribes/troop/utility/TroopFactory.java
@@ -1,10 +1,0 @@
-package rueppellii.backend2.tribes.troop.utility;
-
-import rueppellii.backend2.tribes.troop.persistence.model.Troop;
-
-public class TroopFactory {
-
-    public static Troop makeTroop(){
-        return new Troop();
-    }
-}
